+++ app/Http/Controllers/CRUDController.php
@@ -10,14 +10,32 @@ use Illuminate\Foundation\Http\FormRequest;
 
 class CRUDController extends Controller
 {
+    /**
+     * crudService
+     * @var 
+     */
     protected $crudService;
+    /**
+     * formRequest
+     * @var 
+     */
     protected $formRequest;
 
+    /**
+     * Instantion of $crudServive property and dependency injection of CRUDService
+     * @param \App\Services\CRUDService $crudService
+     */
     public function __construct(CRUDService $crudService)
     {
         $this->crudService = $crudService;
     }
 
+    /**
+     * CRUD method to create an entity
+     * @param \Illuminate\Foundation\Http\FormRequest $request
+     * @param mixed $objectType
+     * @return \Illuminate\Http\JsonResponse|mixed
+     */
     public function create(FormRequest $request, $objectType)
     {
         $data = $this->crudService->store($request->validated());
@@ -27,6 +45,13 @@ class CRUDController extends Controller
         ], 201);
     }
 
+    /**
+     * CRUD method to perform a full update on an entity
+     * @param \Illuminate\Foundation\Http\FormRequest $request
+     * @param \Illuminate\Database\Eloquent\Model $model
+     * @param mixed $objectType
+     * @return \Illuminate\Http\JsonResponse|mixed
+     */
     public function updateAll(FormRequest $request, Model $model, $objectType)
     {
         $model->update($request->validated());
@@ -37,6 +62,13 @@ class CRUDController extends Controller
         ]);
     }
 
+    /**
+     * CRUD method to perform a partial update on an entity
+     * @param \Illuminate\Foundation\Http\FormRequest $request
+     * @param mixed $id
+     * @param mixed $objectType
+     * @return \Illuminate\Http\JsonResponse|mixed
+     */
     public function patch(FormRequest $request, $id, $objectType)
     {
         $data = $this->crudService->edit($id, $request->validated());
@@ -46,6 +78,12 @@ class CRUDController extends Controller
         ]);
     }
 
+    /**
+     * CRUD method to destroy an entity
+     * @param \Illuminate\Database\Eloquent\Model $model
+     * @param mixed $objectType
+     * @return \Illuminate\Http\JsonResponse|mixed
+     */
     public function destroy(Model $model, $objectType)
     {
         $this->crudService->destroy($model);
@@ -54,11 +92,20 @@ class CRUDController extends Controller
         ]);
     }
 
+    /**
+     * CRUD controller method to retrieve all available entities
+     * @return \Illuminate\Database\Eloquent\Collection
+     */
     public function all()
     {
         return $this->crudService->getAll();
     }
 
+    /**
+     * CRUD controller method to retrieve a single entity by id
+     * @param mixed $id
+     * @return mixed
+     */
     public function single($id)
     {
         return $this->crudService->getById($id);
