+++ app/Services/ComponentService.php
@@ -5,12 +5,16 @@ namespace App\Services;
 use App\Models\Component;
 use App\Services\CRUDService;
 
+
 /**
- * Class ComponentService.
+ * Summary of ComponentService
  */
 class ComponentService extends CRUDService
 {
 
+    /**
+     * Call to parent constructor and dependency injection
+     */
     public function __construct()
     {
         parent::__construct(new Component());
