+++ app/Services/GradeService.php
@@ -4,11 +4,12 @@ namespace App\Services;
 
 use App\Models\Grade;
 
-/**
- * Class GradeService.
- */
+
 class GradeService extends CRUDService
 {
+    /**
+     * Call to parent constructor and dependency injection
+     */
     public function __construct()
     {
         parent::__construct(new Grade());
