+++ app/Services/InspectionService.php
@@ -7,16 +7,20 @@ use App\Services\CRUDService;
 
 new Inspection;
 
-/**
- * Class InspectionService.
- */
 class InspectionService extends CRUDService
 {
+    /**
+     * Call to parent constructor and dependency injection
+     */
     public function __construct()
     {
         parent::__construct(new Inspection());
     }
 
+    /**
+     * Method to retrieve all inspections, along with their relationships
+     * @return \Illuminate\Http\JsonResponse|mixed
+     */
     public function getAllWithRelationships()
     {
         $inspectionsWithTurbines = Inspection::with('turbine', 'turbine.components')->get();
@@ -26,6 +30,11 @@ class InspectionService extends CRUDService
 
 
 
+    /**
+     * Method to retrieve an inspection, along with its reltionships
+     * @param mixed $id
+     * @return \Illuminate\Http\JsonResponse|mixed
+     */
     public function getAnInspectionWithRelationships($id)
     {
         $inspection = Inspection::with('turbine', 'turbine.components')->where('id', $id)->get();
