+++ app/Services/TurbineComponentService.php
@@ -3,11 +3,12 @@
 namespace App\Services;
 use App\Models\TurbineComponent;
 
-/**
- * Class TurbineComponentService.
- */
+
 class TurbineComponentService extends CRUDService
 {
+    /**
+     * Call to parent constructor and dependency injection
+     */
     public function __construct()
     {
         parent::__construct(new TurbineComponent());
